KA-137 -Rename Loader files to ListOrSpinner
